backend/paragraphs/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
beautifulsoup4==4.13.3
    # via bs4
bs4==0.0.2
    # via -r requirements.in
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
cobble==0.1.4
    # via mammoth
et-xmlfile==2.0.0
    # via openpyxl
idna==3.10
    # via requests
lxml==5.3.1
    # via
    #   python-docx
    #   python-pptx
mammoth==1.9.0
    # via -r requirements.in
openpyxl==3.1.5
    # via -r requirements.in
pillow==11.1.0
    # via python-pptx
pypdf2==3.0.1
    # via -r requirements.in
python-docx==1.1.2
    # via -r requirements.in
python-pptx==0.6.23
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
soupsieve==2.6
    # via beautifulsoup4
striprtf==0.0.28
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   beautifulsoup4
    #   pypdf2
    #   python-docx
urllib3==2.3.0
    # via requests
xlsxwriter==3.2.2
    # via python-pptx
